ajuste

+++ dev-superior-algoritmo-e-logica-de-programacao/c/variaveis.c
@@ -4,7 +4,7 @@
 int main() {
 
     int idade;
-    float salario, altura;
+    double salario, altura;
     char genero;
     char nome[50];
